requirements.txt
flask
groq
python-dotenv
google.genai
